http: add credentials parser
Add a parser for credentials in HTTP Authorization header.

For example, if you have an Authorization header:

    Authorization: Custom k1=v1, k2=v2

credentials are "Custom k1=v1, k2=v2" and the parser parses them as
follows:

    auth-scheme: Custom
    auth-params: k1: v1
                 k2: v2

This parser allows credentials which contain both of token68 and auth
params, even if the HTTP Authorization specification [RFC7235] forbids
it, to support erroneous implementations. For example:

Credentials:

    Custom abc, k1=v1

The parsed result:

    auth-scheme: Custom
    token: abc
    auth-params: k1: v1
                 k2: v2

[RFC7235]: https://tools.ietf.org/html/rfc7235